refactor: move PokemonList to separate folder and create corresponding css file
