+++ src/components/Home.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import Search from './Search/Search';
-import PokemonList from './PokemonList';
+import PokemonList from './PokemonList/PokemonList';
 import Pagination from './Pagination/Pagination';
 import PokemonDetails from '../components/PokemonDetails';
